288884 - Consolidated tests
